+++ CesiumGeometry/include/CesiumGeometry/Transforms.h
@@ -2,14 +2,15 @@
 
 #include "Library.h"
 
-#include <glm/mat3x3.hpp>
+#include <glm/fwd.hpp>
+#include <glm/mat4x4.hpp>
 
 namespace CesiumGeometry {
 
 /**
- * @brief Coordinate system conversion matrices
+ * @brief Coordinate system matrix constructions helpers.
  */
-struct CESIUMGEOMETRY_API AxisTransforms final {
+struct CESIUMGEOMETRY_API Transforms final {
 
   /**
    * @brief A matrix to convert from y-up to z-up orientation,
@@ -46,6 +47,20 @@ struct CESIUMGEOMETRY_API AxisTransforms final {
    * by rotating about -PI/2 around the z-axis
    */
   static const glm::dmat4 Y_UP_TO_X_UP;
+
+  /**
+   * @brief Constructs a translation-rotation-scale matrix, equivalent to
+   * `translation * rotation * scale`. So if a vector is multiplied with the
+   * resulting matrix, it will be first scaled, then rotated, then translated.
+   *
+   * @param translation The translation.
+   * @param rotation The rotation.
+   * @param scale The scale.
+   */
+  static glm::dmat4 translationRotationScale(
+      const glm::dvec3& translation,
+      const glm::dquat& rotation,
+      const glm::dvec3& scale);
 };
 
 } // namespace CesiumGeometry